refactor cosmos setup
