+++ source/Energinet.DataHub.MessageArchive.EntryPoint/BlobServices/BlobReader.cs
@@ -39,7 +39,7 @@ namespace Energinet.DataHub.MessageArchive.EntryPoint.BlobServices
 
         public async Task<List<BlobItemData>> GetBlobsReadyForProcessingAsync()
         {
-            // can be paged
+            //TODO can be paged
             var blobsToProcess = _blobContainerClient
                 .GetBlobsAsync(BlobTraits.All).ConfigureAwait(false);
