+++ source/Energinet.DataHub.MessageArchive.EntryPoint/local.settings.sample.json
@@ -5,8 +5,6 @@
     "STORAGE_MESSAGE_ARCHIVE_CONTAINER_NAME": "<Setting-Value>",
     "STORAGE_MESSAGE_ARCHIVE_PROCESSED_CONTAINER_NAME": "<Setting-Value>",
     "COSMOS_MESSAGE_ARCHIVE_CONNECTION_STRING": "<Setting-Value>",
-    "COSMOS_MESSAGE_ARCHIVE_DATABASE_ID": "<Setting-Value>",
-    "COSMOS_MESSAGE_ARCHIVE_CONTAINER_NAME": "<Setting-Value>",
     "AzureWebJobsStorage": "UseDevelopmentStorage=true",
     "FUNCTIONS_WORKER_RUNTIME": "dotnet-isolated",
     "APPINSIGHTS_INSTRUMENTATIONKEY": "",
